requirements.txt
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distlib==0.3.7
docopt==0.6.2
filelock==3.13.1
Flask==3.0.2
google-api-core==2.17.1
google-auth==2.28.1
googleapis-common-protos==1.62.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
pipreqs==0.4.13
platformdirs==4.1.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dateutil==2.8.2
pytz==2024.1
rsa==4.9
six==1.16.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.25.0
Werkzeug==3.0.1
yarg==0.1.9
